rename models to controllers
